+++ src/utils/type_list.hpp
@@ -126,7 +126,6 @@ auto GetNames() {
   return names;
 }
 
-//<<<<<<< HEAD
 namespace impl {
 template <class N, class T>
 struct ListOfType {
@@ -148,14 +147,35 @@ struct ListOfType<std::integral_constant<std::size_t, 1>, T> {
 template <std::size_t N, class T>
 using list_of_type_t =
     typename impl::ListOfType<std::integral_constant<std::size_t, N>, T>::type;
-//=======
 template <class>
 struct isTypeList : public std::false_type {};
 
 template <class... Ts>
 struct isTypeList<TypeList<Ts...>> : public std::true_type {};
 
-//>>>>>>> develop
+namespace impl {
+template <class Tuple>
+struct TupleToTypeList;
+
+template <class... Ts>
+struct TupleToTypeList<std::tuple<Ts...>> {
+  using type = TypeList<Ts...>;
+};
+
+template <template <class> class Pred, class... Ts>
+auto FilterTypeList(TypeList<Ts...>) {
+  return std::tuple_cat(
+      std::conditional_t<Pred<Ts>::value, std::tuple<Ts>, std::tuple<>>{}...);
+}
+
+} // namespace impl
+
+template <class Tuple>
+using tuple_to_type_list_t = typename impl::TupleToTypeList<Tuple>::type;
+
+template <class TL, template <class> class Pred>
+using filter_type_list_t =
+    tuple_to_type_list_t<decltype(impl::FilterTypeList<Pred>(std::declval<TL>()))>;
 
 //----------------------------------------------------------------------------------------
 // Utility to convert TypeList to std::variant